% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_numbers.R
\name{f2si}
\alias{f2si}
\title{Transform a numerical value in a readable text string with K, M, G notations}
\usage{
f2si(number, rounding = F, digits = ifelse(rounding, NA, 3))
}
\arguments{
\item{number}{the number you want to transform}

\item{rounding}{TRUE will round the value (no decimal), default is FALSE}

\item{digits}{exact number of digits that will be displayed, default is 3}
}
\value{
this function returns a string with SI format
}
\description{
Transform a numerical value in a readable text string with K, M, G notations
}
\author{
Alban Besnard [aut]
}
