% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblio_network.R
\name{makeRelations}
\alias{makeRelations}
\title{Getting relations between authors}
\usage{
makeRelations(bibtex)
}
\arguments{
\item{bibtex}{a bibtex object containing the bibliography you want to work with}
}
\value{
this function will give you a data.frame of relations (one line per relation)
}
\description{
Useful to plot a network
From a bibtex object (RefManager),
}
\author{
Alban Besnard [aut]
}
