% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblio_network.R
\name{makeVertices}
\alias{makeVertices}
\title{Creating groups}
\usage{
makeVertices(relations)
}
\arguments{
\item{relations}{data.frame as created by the makeRelation function}
}
\value{
a data.frame indicating the frequency of each author and his calculated group
}
\description{
Calculating and assigning groups of researchers based on betweeness centrality.
}
\author{
Alban Besnard [aut]
}
