% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblio_network.R
\name{updateRelations}
\alias{updateRelations}
\title{Adding indexes}
\usage{
updateRelations(relations, vertices)
}
\arguments{
\item{relations}{data.frame as created by the makeRelation function}

\item{vertices}{data.frame as created by the makeVertices function}
}
\value{
a data.frame similar to relations updated with relation indexes
}
\description{
Adding indexes
}
\author{
Alban Besnard [aut]
}
